Update button styles
