+++ .storybook/main.js
@@ -1,3 +1,4 @@
 module.exports = {
+  addons: ['@storybook/addon-knobs/register'],
   stories: ['../stories/**/*.stories.js'],
 };
