+++ package.json
@@ -38,6 +38,7 @@
     "@figma-export/cli": "^1.2.0",
     "@figma-export/output-components-as-svg": "^1.2.0",
     "@figma-export/transform-svg-with-svgo": "^1.1.1",
+    "@storybook/addon-knobs": "^5.3.17",
     "@storybook/html": "^5.3.17",
     "@types/jest": "^25.1.4",
     "@types/node": "^13.9.3",
