+++ src/components/button.scss
@@ -1,152 +1,130 @@
 @use "../design-tokens/color";
+@use "../design-tokens/gradient";
+@use "../design-tokens/shadow";
+@use "../design-tokens/typography";
 
 .Manifold__Button {
-  align-items: center;
+  appearance: none;
   background: color.$color-white;
-  border-color: hsl(210, 5%, 78%);
+  border-color: color.$color-borderGray;
   border-radius: 4px;
   border-style: solid;
   border-width: 1px;
-  box-shadow: 0 2px 8px rgba(color.$color-black, 0.08);
+  box-shadow: shadow.$shadow-near;
   box-sizing: border-box;
   color: color.$color-grayDarker;
   cursor: pointer;
-  display: inline-flex;
-  font-family: var(--f-body);
-  font-size: 16px;
+  display: inline-block;
+  font-family: typography.$typography-body-fontFamily;
+  font-size: typography.$typography-body-fontSize;
   font-weight: 500;
-  height: 2.5rem;
-  justify-content: center;
-  line-height: 2.5;
-  outline: none;
-  padding: 0 1.25em;
-  text-align: center;
+  line-height: 1;
+  padding: 10.5px 21px;
+  position: relative;
   text-decoration: none;
   transition-duration: 150ms;
   transition-property: background-color, border-color, color, box-shadow, filter;
   transition-timing-function: linear;
   white-space: nowrap;
 
+  &::before {
+    border: 2px solid color.$color-blue;
+    border-radius: 8px;
+    bottom: -5px;
+    content: '';
+    left: -5px;
+    pointer-events: none;
+    position: absolute;
+    right: -5px;
+    top: -5px;
+    opacity: 0;
+    transform: scale(0.95, 0.9); // x scale is more noticable
+    transition: opacity 150ms linear, transform 150ms cubic-bezier(0, 0.5, 0.5, 1);
+  }
+
+  &::-moz-focus-inner {
+    border: 0;
+  }
+
   &:focus,
   &:hover {
-    filter: saturate(130%);
+    border-color: #8f8f95;
+  }
+
+  &:focus {
+    outline: none;
+
+    &::before {
+      opacity: 1;
+      transform: scale(1);
+    }
   }
 
   &:active {
-    filter: brightness(90%);
     transform: translateY(1px);
   }
 
-  & > svg {
-    margin-inline-start: -0.0625em; /* icon fix */
-  }
-
   &[data-color='black'] {
-    background: color.$color-black;
-    border: transparent;
-    color: rgba(color.$color-white, 0.95);
-    &:active {
-      color: rgba(color.$color-white, 0.9);
-    }
-    &:focus,
-    &:hover {
-      background: rgba(color.$color-black, 0.9);
-    }
-  }
+    background: color.$color-grayDarker;
+    border-color: transparent;
+    color: color.$color-white;
 
-  &[data-color='gray'] {
-    background: hsla(var(--c-grayDarker), 0.1);
-    border-color: hsla(var(--c-grayDarker), 0.4);
-    box-shadow: none;
-    color: hsla(var(--c-gray));
-    &:active {
-      background: hsla(var(--c-gray), 0.05);
-      color: hsla(var(--c-gray), 0.8);
-    }
-    &:focus,
     &:hover {
-      color: hsl(var(--c-gray));
-    }
-  }
-
-  &[data-color='orange'] {
-    background: var(--g-orange);
-    border-style: none;
-    box-shadow: none;
-    color: hsl(var(--c-white));
-    &:active {
-      color: hsla(var(--c-white), 0.8);
+      background: color.$color-black;
     }
   }
 
   &[data-color='brand'] {
-    background: var(--g-brand);
-    border-style: none;
-    box-shadow: none;
-    color: hsl(var(--c-white));
-    &:active {
-      color: hsla(var(--c-white), 0.8);
+    background: gradient.$gradient-brand;
+    border-color: transparent;
+    color: color.$color-white;
+
+    &:hover {
+      filter: saturate(150%);
     }
   }
 
-  &[data-color='red'] {
-    background: hsl(var(--c-red));
-    border-style: none;
-    box-shadow: none;
-    color: hsl(var(--c-white));
-    &:active {
-      color: hsla(var(--c-red), 0.8);
-    }
-  }
-
-  &[data-color='red-knockout'] {
-    background: hsla(var(--c-white));
+  &[data-color='danger'] {
+    background: color.$color-white;
+    color: color.$color-red;
     border-color: currentColor;
-    color: hsl(var(--c-red));
-    box-shadow: none;
-    &:active {
-      color: hsla(var(--c-red), 0.8);
+
+    &:hover {
+      background: #fef7f9;
     }
   }
 
-  &[data-color='white'] {
-    &:active {
-      border-color: hsla(var(c--black), 0.4);
-      box-shadow: 0 2px 8px hsla(var(--c-black), 0);
-      color: hsl(var(--c-gray));
-    }
+  &:disabled {
+    background: color.$color-grayLighter;
+    border-color: color.$color-borderGray;
+    box-shadow: none;
+    color: color.$color-gray;
+    cursor: not-allowed;
+    opacity: 0.8;
+
     &:focus,
     &:hover {
-      border-color: hsla(var(c--black), 0.4);
-      color: hsl(var(--c-black));
+      background: color.$color-grayLighter;
+      border-color: color.$color-borderGray;
+      box-shadow: none;
+      color: color.$color-gray;
+    }
+
+    &:focus::before {
+      opacity: 0;
+    }
+
+    &:active {
+      background: color.$color-grayLighter;
+      border-color: color.$color-borderGray;
+      box-shadow: none;
+      color: color.$color-gray;
+      transform: none;
     }
   }
 
   &[data-size='small'] {
-    font-size: 12px;
-    height: 2rem;
-    line-height: 1;
-    padding: 0 1rem;
-  }
-
-  &:disabled {
-    background: hsla(var(--c-black), 0.05);
-    border-color: transparent;
-    box-shadow: none;
-    color: hsla(var(--c-black), 0.6);
-    cursor: not-allowed;
-    opacity: 0.8;
-    &:focus,
-    &:hover {
-      background: hsla(var(--c-black), 0.05);
-      box-shadow: none;
-      color: hsla(var(--c-black), 0.6);
-    }
-    &:active {
-      background: hsla(var(--c-black), 0.05);
-      box-shadow: none;
-      color: hsla(var(--c-black), 0.6);
-      transform: none;
-    }
+    font-size: 13px;
+    padding: 8px 14px;
   }
 }
